Split editor vertex mode helpers
