+++ CMakeLists.txt
@@ -175,6 +175,7 @@ target_sources(
         src/game_data_editor_selection_output.c
         src/game_data_editor_selection_helpers.c
         src/game_data_editor_selection_state.c
+        src/game_data_editor_vertex_mode.c
         src/game_data_editor_selection_queries.c
         src/game_data_editor_debug_primitives.c
         src/game_data_editor_grid_widget.c
